Commissions and rate value type to DECIMAL

+++ src/main/resources/db/migration/V3_0__create_commissions_table.sql
@@ -3,6 +3,6 @@ create TABLE commissions
 (
     from_currency    VARCHAR(60) NOT NULL,
     to_currency      VARCHAR(60) NOT NULL,
-    value   DOUBLE PRECISION NOT NULL,
+    value            DECIMAL(19, 4)  NOT NULL,
     PRIMARY KEY(from_currency, to_currency)
 );
