+++ src/main/resources/db/migration/V4_0__create_exchange_rates_table.sql
@@ -3,6 +3,6 @@ CREATE TABLE exchange_rates
 (
     from_currency    VARCHAR(60) NOT NULL,
     to_currency      VARCHAR(60) NOT NULL,
-    rate   DOUBLE PRECISION NOT NULL,
+    rate             DECIMAL(19, 4) NOT NULL,
     PRIMARY KEY(from_currency, to_currency)
 );
